refactored HealthCheck - changed healthcheck score type from Int to Short

+++ src/main/scala/com/azaptree/application/healthcheck/healthcheck.scala
@@ -9,14 +9,14 @@ case class HealthCheck(
     description: Option[String] = None,
     enabled: Boolean = true,
     /** Used to indicate the importance of a healthcheck relative to other healthchecks */
-    importanceLevel: Short = 0,
+    importanceLevel: Byte = 0,
     greenRange: GreenHeathCheckIndicatorThreshold = GreenHeathCheckIndicatorThreshold(90),
     yellowRange: YellowHeathCheckIndicatorThreshold = YellowHeathCheckIndicatorThreshold(75),
     redRange: RedHeathCheckIndicatorThreshold = RedHeathCheckIndicatorThreshold(0),
     config: Option[Config] = None,
     schedule: Option[TaskSchedule] = None)(displayName: String = s"$group:$name") {
 
-  def computeHealthCheckIndicator(healthScore: Int): HealthCheckIndicator = {
+  def computeHealthCheckIndicator(healthScore: Short): HealthCheckIndicator = {
     if (healthScore >= greenRange.minScore) {
       GREEN
     } else if (healthScore >= yellowRange.minScore) {
@@ -30,18 +30,18 @@ case class HealthCheck(
 
 sealed trait HeathCheckIndicatorThreshold {
   val indicator: HealthCheckIndicator
-  val minScore: Int
+  val minScore: Short
 }
 
-case class GreenHeathCheckIndicatorThreshold(minScore: Int) extends HeathCheckIndicatorThreshold {
+case class GreenHeathCheckIndicatorThreshold(minScore: Short) extends HeathCheckIndicatorThreshold {
   val indicator = GREEN
 }
 
-case class YellowHeathCheckIndicatorThreshold(minScore: Int) extends HeathCheckIndicatorThreshold {
+case class YellowHeathCheckIndicatorThreshold(minScore: Short) extends HeathCheckIndicatorThreshold {
   val indicator = YELLOW
 }
 
-case class RedHeathCheckIndicatorThreshold(minScore: Int) extends HeathCheckIndicatorThreshold {
+case class RedHeathCheckIndicatorThreshold(minScore: Short) extends HeathCheckIndicatorThreshold {
   val indicator = RED
 }
 
@@ -57,7 +57,7 @@ case class HealthCheckResult(
   executionStart: Long,
   /** when the healthcheck execution ended in Epoch time */
   executionEnd: Long,
-  healthScore: Int,
+  healthScore: Short,
   healthCheckIndicator: HealthCheckIndicator,
   info: Option[String] = None,
   exceptionStackTrace: Option[String] = None)
